requirements.txt
flask
flask-sqlalchemy
requests
python-dotenv
gunicorn
